+++ src/Query/ActivityBuilder.php
@@ -139,6 +139,30 @@ class ActivityBuilder extends Builder
     }
 
     /**
+     * Returns the number of records that match the query. This number is from the API
+     * and might be lower than the actual number of records when queried with get().
+     *
+     * @throws OnOfficeException
+     */
+    public function count(): int
+    {
+        $request = new OnOfficeRequest(
+            OnOfficeAction::Read,
+            OnOfficeResourceType::Activity,
+            parameters: [
+                ...$this->prepareEstateOrAddressParameters(),
+                OnOfficeService::DATA => [],
+                OnOfficeService::FILTER => $this->getFilters(),
+                OnOfficeService::LISTLIMIT => 1,
+                ...$this->customParameters,
+            ]
+        );
+
+        return $this->requestApi($request)
+            ->json('response.results.0.data.meta.cntabsolute', 0);
+    }
+
+    /**
      * @deprecated Use estateId() instead
      */
     public function estate(): static